requirements.txt
certifi            2024.6.2
charset-normalizer 3.3.2
idna               3.7
pi                 0.1.2
pillow             10.3.0
pip                24.0
PyYAML             6.0.1
requests           2.32.3
setuptools         70.0.0
urllib3            2.2.1
wheel              0.43.0